requirements.txt
streamlit
google-genai
PyPDF2
streamlit-agraph
plotly
numpy
psycopg2-binary
python-dotenv
toml
werkzeug
